front end for 4 sliders
