+++ app/assets/stylesheets/components/_filter.scss
@@ -1,0 +1,44 @@
+#filter-pannel{
+  height:200px;
+  margin:3px;
+  display:flex;
+  justify-content: space-around;
+  align-items: center;
+  padding-right: 15px;
+  padding-left: 15px;
+}
+
+.slider-wrapper{
+  width:25px;
+  height:190px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  font-size:22px;
+  position: relative;
+}
+
+.slider{
+  -webkit-appearance: none;
+  height: 3px;
+  width: 110px;
+  outline: none;
+  background: darkgrey;
+  transform: rotate(90deg);
+  border-radius: 1px;
+  position: absolute;
+  right: -41px;
+  top:90px;
+}
+
+.slider::-webkit-slider-thumb {
+  -webkit-appearance: none; /* Override default look */
+  appearance: none;
+  width: 5px; /* Set a specific slider handle width */
+  height: 20px; /* Slider handle height */
+  background: black; /* Green background */
+  border-radius: 3px;
+  cursor: pointer;
+}
+
+
