+++ app/assets/stylesheets/components/_index.scss
@@ -2,3 +2,4 @@
 @import "alert";
 @import "avatar";
 @import "navbar";
+@import "filter";
