+++ app/views/playlists/index.html.erb
@@ -18,7 +18,9 @@
 <%end %>
 </div>
 
-<div class="spotify-embed"><iframe src="https://embed.spotify.com/?uri=spotify:playlist:4G4urZbsLroiZ6Inr4CJNC?si=X57TrMaYSo--hdUiPasIxQ" height=auto frameborder="0" allowtransparency="true" allow="encrypted-media"></iframe>
+<div class="spotify-embed">
+   <%= render 'shared/filter'%>
+  <iframe src="https://embed.spotify.com/?uri=spotify:playlist:4G4urZbsLroiZ6Inr4CJNC?si=X57TrMaYSo--hdUiPasIxQ" height=auto frameborder="0" allowtransparency="true" allow="encrypted-media"></iframe>
 </div>
