+++ app/views/shared/_filter.html.erb
@@ -1,0 +1,9 @@
+
+<div id = "filter-pannel">
+
+<div class = "slider-wrapper">🕺<input type="range" min="1" max="5" value="3" name="danceability" id="danceability-slider" class="slider">🧘‍♀️</div>
+<div class = "slider-wrapper"> 🗣<input type="range" min="1" max="5" value="3" name="vocability" id="vocability-slider" class="slider"> 🤐</div>
+<div class = "slider-wrapper">☀️<input type="range" min="1" max="5" value="3" name="acousticality" id="acousticality-slider" class="slider">❄️</div>
+<div class = "slider-wrapper"> 🤖<input type="range" min="1" max="5" value="3" name="summerability" id="summerability-slider" class="slider"> 🎻</div>
+</div>
+
