+++ app/views/shared/_navbar.html.erb
@@ -3,7 +3,6 @@
     <%= image_tag "white tape.jpg" %>
     <% end %>
 
-    <input type="range" min="0" max="2" value="1" name="danceability" id="danceability-slider" class="slider">
 
   <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
     <span class="navbar-toggler-icon"></span>
